+++ src/AhkSymbol/TAhkSymbolIn.ts
@@ -1,12 +1,12 @@
 import type { CAhkClass, CAhkClassGetSet, CAhkClassInstanceVar } from './CAhkClass';
 import type { CAhkFunc } from './CAhkFunc';
+import type { CAhkHotString } from './CAhkHotString';
 import type {
     CAhkComment,
     CAhkHotKeys,
-    CAhkHotString,
-    TLineClass,
 } from './CAhkLine';
 import type { CAhkCase, CAhkDefault, CAhkSwitch } from './CAhkSwitch';
+import type { TLineClass } from './TLineClass';
 
 export type TAhkSymbol =
     | CAhkCase