+++ src/AhkSymbol/TLineClass.ts
@@ -1,0 +1,8 @@
+import type { CAhkInclude } from './CAhkInclude';
+import type { CAhkComment, CAhkDirectives, CAhkLabel } from './CAhkLine';
+
+export type TLineClass =
+    | CAhkComment
+    | CAhkDirectives
+    | CAhkInclude
+    | CAhkLabel;